feat: 重构McpServerConfigProcessor为McpServerConfigurationLoader，优化MCP服务器配置处理逻辑
